fix(migrations): add baseline schema so migration:run works from scratch
Add a baseline schema migration (lowest timestamp) reproducing the schema
TypeORM's synchronize previously created in development, so migration:run
succeeds on a fresh Postgres database and the incremental migrations apply
cleanly on top.

Also fixes several pre-existing blockers that surfaced once migrations ran
from scratch:
- datasource now loads entities so schema-aware operations compare against
  real entity metadata (and typeorm_metadata is created for generated columns)
- UserChallenge index referenced non-existent property names
- onboarding entities had duplicate class-level indexes colliding with
  property-level ones
- CourseStatus enum extracted to its own module to break a circular import
- reencrypt-oauth-provider-tokens no longer demands ENCRYPTION_SECRET on
  empty databases
- add-unique-course-version-constraint referenced non-existent columns
- add-paused-subscription-status targeted the wrong enum type name
- fix-invoice-number-sequence opened a second connection that couldn't see
  uncommitted tables, and setval(..., 0) failed on empty tables

The drift check now compares against scripts/known-drift.txt (pre-existing
entity/migration naming drift, tracked in issue #1194) and only fails when
new drift is introduced. Closes #1193
