+++ package.json
@@ -46,7 +46,7 @@
     "migration:run": "npx typeorm-ts-node-commonjs migration:run -d src/config/datasource.ts",
     "migration:revert": "npx typeorm-ts-node-commonjs migration:revert -d src/config/datasource.ts",
     "migration:generate": "npx typeorm-ts-node-commonjs migration:generate -d src/config/datasource.ts",
-    "migration:generate:check": "BEFORE=$(find src/migrations -maxdepth 1 -name '[0-9]*.ts' | wc -l) && npx typeorm-ts-node-commonjs migration:generate -d src/config/datasource.ts src/migrations/drift-check-$(date +%s) && AFTER=$(find src/migrations -maxdepth 1 -name '[0-9]*.ts' | wc -l) && find src/migrations -maxdepth 1 -name '*drift-check*' -delete && if [ \"$AFTER\" -gt \"$BEFORE\" ]; then echo \"ERROR: Schema drift detected - model changes exist without a corresponding migration. Run npm run migration:generate to create the missing migration file.\" && exit 1; fi",
+    "migration:generate:check": "bash scripts/drift-check.sh",
     "docs:generate": "node scripts/generate-api-docs.js",
     "docs:validate": "node scripts/validate-openapi.js",
     "docs:check": "npm run docs:generate && npm run docs:generate:examples && git diff --exit-code -- openapi-spec.json docs/api/openapi-spec.json docs/api/examples.md docs/site docs/examples",
