+++ src/courses/entities/course-status.enum.ts
@@ -1,0 +1,9 @@
+/** Lifecycle states a course can be in. */
+export enum CourseStatus {
+  DRAFT = 'draft',
+  PENDING_REVIEW = 'pending_review',
+  CHANGES_REQUESTED = 'changes_requested',
+  PUBLISHED = 'published',
+  REJECTED = 'rejected',
+  ARCHIVED = 'archived',
+}
